Move MapGen and MapHelpers units into map_generator folder
