+++ src/openrct2-ui/windows/MapGen.cpp
@@ -25,7 +25,7 @@
 #include <openrct2/sprites.h>
 #include <openrct2/windows/Intent.h>
 #include <openrct2/world/Map.h>
-#include <openrct2/world/MapGen.h>
+#include <openrct2/world/map_generator/MapGen.h>
 
 namespace OpenRCT2::Ui::Windows
 {
