+++ src/openrct2/world/map_generator/MapGen.cpp
@@ -9,29 +9,29 @@
 
 #include "MapGen.h"
 
-#include "../Context.h"
-#include "../Diagnostic.h"
-#include "../Game.h"
-#include "../GameState.h"
-#include "../core/Guard.hpp"
-#include "../core/Imaging.h"
-#include "../core/String.hpp"
-#include "../localisation/StringIds.h"
-#include "../object/ObjectEntryManager.h"
-#include "../object/ObjectList.h"
-#include "../object/ObjectManager.h"
-#include "../object/SmallSceneryEntry.h"
-#include "../object/TerrainEdgeObject.h"
-#include "../object/TerrainSurfaceObject.h"
-#include "../platform/Platform.h"
-#include "../util/Util.h"
-#include "../world/tile_element/Slope.h"
-#include "../world/tile_element/SmallSceneryElement.h"
-#include "../world/tile_element/SurfaceElement.h"
-#include "../world/tile_element/TileElement.h"
-#include "Map.h"
+#include "../../Context.h"
+#include "../../Diagnostic.h"
+#include "../../Game.h"
+#include "../../GameState.h"
+#include "../../core/Guard.hpp"
+#include "../../core/Imaging.h"
+#include "../../core/String.hpp"
+#include "../../localisation/StringIds.h"
+#include "../../object/ObjectEntryManager.h"
+#include "../../object/ObjectList.h"
+#include "../../object/ObjectManager.h"
+#include "../../object/SmallSceneryEntry.h"
+#include "../../object/TerrainEdgeObject.h"
+#include "../../object/TerrainSurfaceObject.h"
+#include "../../platform/Platform.h"
+#include "../../util/Util.h"
+#include "../../world/tile_element/Slope.h"
+#include "../../world/tile_element/SmallSceneryElement.h"
+#include "../../world/tile_element/SurfaceElement.h"
+#include "../../world/tile_element/TileElement.h"
+#include "../Map.h"
+#include "../Scenery.h"
 #include "MapHelpers.h"
-#include "Scenery.h"
 
 #include <cmath>
 #include <cstring>
