+++ src/openrct2/world/map_generator/MapHelpers.cpp
@@ -9,9 +9,9 @@
 
 #include "MapHelpers.h"
 
-#include "../world/tile_element/Slope.h"
-#include "../world/tile_element/SurfaceElement.h"
-#include "Map.h"
+#include "../../world/tile_element/Slope.h"
+#include "../../world/tile_element/SurfaceElement.h"
+#include "../Map.h"
 
 #include <algorithm>
